code_genius/codebase_genius/BE/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
asyncer==0.0.8
attrs==25.4.0
broadcaster==0.3.1
byllm==0.4.5
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecs-logging==2.2.0
email_validator==2.2.0
fakeredis==2.27.0
fastapi==0.115.11
fastapi-sso==0.17.0
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.38.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
importlib_metadata==8.7.0
jac-cloud==0.2.10
jaclang==0.8.10
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.79.3
loguru==0.7.3
MarkupSafe==3.0.3
montydb==2.5.3
motor==3.7.0
multidict==6.7.0
oauthlib==3.3.1
openai==2.7.1
orjson==3.10.15
packaging==25.0
passlib==1.7.4
pillow==10.4.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv==1.2.1
python-http-client==3.3.7
python-multipart==0.0.20
PyYAML==6.0.3
redis==5.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
sendgrid==6.11.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
starkbank-ecdsa==2.2.0
starlette==0.46.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
tree-sitter==0.25.2
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
